Move bulk of Writer into DocxWriter
